Download detector data with timings
You can use the form to download detector data and trace timings are
calculated on the fly.  This is, however, a pretty slow process.

(I suggest we do develop a script which goes over past events and
calculates and then stores the timings.  Maybe we should lock event tables
past a certain age (a week?) and only manually allow inclusion of 'old'
data.)