Generated csv files contain tabs and #
The csv files which were generated by clicking on 'source' contained
semi-colons as column delimiters, to make some versions of Excel
automatically load the file correctly.  However, the proliferation of
different versions and languages of Excel make this not useful in
practice.

Furthermore, it is hard to read.  Even more importantly, numpy.genfromtxt
fails to understand this format.

Finally, the file contains tabs as delimiters and hashes for comment lines.