ProcessEventsFromSource, with test setup
Temporary files are not yet correctly implemented, but the datastore event
tables are processed (duplicates are not yet removed!!) and the results
are stored in an intermediate file.  The events node from this
intermediate file is copied to the ESD.

To test all this between code changes, navigate to the repo root and run::

    $ sh test.sh

This will reset the histograms app (this sometimes fails, just re-run),
sync the database and runs the django cron app.