Split building ESD in two parts
The first part processes events and stores the results in a temporary file
(as previously), but *returns the filename and node path*.  The second
part then takes the filename and node path and copies the node to the ESD
and removes the temporary file.

Splitting the work in two parts enables us to parallelize the processing
while copying all nodes sequentially.