Use multiprocessing to generate ESD data
A worker pool is created (cpu_count() number of processes) to process
events and create the temporary ESD data.  Whenever a worker finishes, the
data is moved by the main process into the ESD.  So now we have
concurrency for the cpu-intensive processing of events, and only one
process which writes to the HDF5 files.