The Multiprocessing Meltdown Restoration Project
Of course, it was too simple.  Testing on MySQL uncovered too much shared
state in the child processes.  In particular, the database connection was
shared.  That is mitigated by calling django.db.close_connection() in the
child processes.  Also, map_async behaves more eratically than
imap_unordered (and waiting for the results).  I do not know why, and it
bothers me.  This commit fixes things and all should be well.