Attempt to kill off active ssh connection
Ansible uses persistent ssh connections to greatly speed up the running of
tasks on remote machines.  However, when adding the vagrant user to the
hisparc group, the subsequent tasks reuse the connection and won't see the
updated group membership.  You need to log out / log in for that.  As I
can't find an option to forcibly renew the ssh connection, I decided to
kill it off from the remote end.