Nicely handle other files in the datastore
There should not be anything else, but just in case handle it.
On a Mac the Finder likes to create .DS_Store files.
These will be found by the os.walk, but do not fit the format for
strptime.
This try/except nicely ignores files with the wrong name format.