Fix Django MySQL connection
Compiling mysql-python uses `mysql_config` which outputs the location of
the socket specified in /etc/my.cnf. So this works. However, using conda,
mysql-python is precompiled and defaults to /tmp/mysql.sock. So Django
cannot connect to the mysql database, unless you specify 127.0.0.1 as the
host. Then it doesn't use sockets.