Reduce datapoints stored for datasets
Up to now all points were stored for the barometer/temperature datasets.
Since these are measured every 5 seconds there are 17280 points per day.
This takes up a lot of space, and if already sliced for display on the
site.
Currently the datasets take 1.4 GB in the publicdb, the histograms only
360 MB.
However, there are far more histograms.
This should reduce the storage size (by factor ~30) and not affect the
display.