Provisioning: incorporate changes from Pique
The hard drives on Pique were replaced with two newer 2 Tb drives
operating in RAID-1.  We reinstalled the system on 2014/4/17 using ks.cfg
from the veewee directory and used ansible to provision the system.
Several changes were made manually (mostly by Arne) and those are
incorporated in this commit.  Note that several passwords and settings can
only be used on the production system and are not committed.