Add migration to load initial generator state
When applying this on the server for the fist time the `last_run`
for both the check and update should be set to their previous value.
No longer need to run the loaddata in the provisioning for the vagrant
host.
Using migrations should ensure that this data is only loaded once.