Add downloading data with a comment line
Always end with a comment line (if download is successful).
If no data matches the query end with a line stating that.
If some data is returned end with 'Finished downloading.' comment.
This can be used to detect failed downloads,
in which case it does not end with a comment if data was returned,
or with an empty comment in case of no data.
The last line of the headers are always empty.