Reorder nagios status codes
Fix for a specifix corner case: we have vu1 and vu2. I shutdown vu1, and
made it inactive in the admin panel. The link to the station status page
is now to vu2. Fine. The status marker on the station list page was red,
however, not green. I fixed that.