Add SELinux module for nginx
Allow nginx to connect via unix socket to uWSGI
Allow nginx to read general files (serve /static and /media)

The latter policy is a bit of a security risk. It would be better
to label the files in /srv/publicdb/static to `httpd_t` or something,
but that would break django `collectreload`. This SELinux stuff is just
too difficult to get right.