Check server authorisation for IPv4 and/or IPv6
The Nikhef network is IPv6, so both tietar and frome connect via IPv6.
As `socket.gethostbyname` is IPv4 only, the authorisation check fails.

This is a reverse-DNS lookup using `socket.gethostbyaddr` which can be
compared to the hostname. It is as (in)secure as the previous check, but
works for IPv4 and IPv6.