Extract kernel info from grub config.
The current CentOS kernel fails to boot on our hardware. CT patched grub to
boot an old one (3.10.0-514.el7.x86_64). As a result this scripts always
flags a reboot, because a newer kernel is availabe. This is a hack to
prevent rebooting our servers at every provisioning run.
This extracts kernel info from the menuentry lines in GRUB config.